Fixed up warnings

+++ Frameworks/File_Extractor/File_Extractor/fex/Binary_Extractor.cpp
@@ -71,7 +71,7 @@ blargg_err_t Binary_Extractor::stat_v()
 	return open_v();
 }
 
-blargg_err_t Binary_Extractor::extract_v( void* p, int n )
+blargg_err_t Binary_Extractor::extract_v( void* p, long n )
 {
 	return arc().read( p, n );
 }
