+++ Frameworks/File_Extractor/File_Extractor/unrar/rawread.hpp
@@ -6,20 +6,20 @@ class RawRead
 private:
 	Array<byte> Data;
 	File *SrcFile;
-	int DataSize;
-	int ReadPos;
+	long DataSize;
+	long ReadPos;
 	friend class Archive;
 public:
 	RawRead(File *SrcFile);
 	void Reset();
-	void Read(int Size);
+	void Read(long Size);
 	void Get(byte &Field);
 	void Get(ushort &Field);
 	void Get(uint &Field);
 	void Get(byte *Field,int Size);
 	uint GetCRC(bool ProcessedOnly);
-	int Size() {return DataSize;}
-    int PaddedSize() {return Data.Size()-DataSize;}
+	long Size() {return DataSize;}
+    long PaddedSize() {return Data.Size()-DataSize;}
 };
 
 #endif
